requirements.txt
pandas
numpy
streamlit
scikit-learn
pymysql
python-dotenv